Updated Borland port of Win32 to compile cleanly and run the demo.

+++ bacnet-stack/bip.c
@@ -217,7 +217,7 @@ uint16_t bip_receive(
     int max;
     struct timeval select_timeout;
     struct sockaddr_in sin = {-1};
-    size_t sin_len = sizeof(sin);
+    int sin_len = sizeof(sin);
 
     /* Make sure the socket is open */
     if (BIP_Socket < 0)
